Rename Language.Haskell.LSP.Test => Language.LSP.Test
