+++ src/Language/LSP/Test/Replay.hs
@@ -1,6 +1,6 @@
 -- | A testing tool for replaying captured client logs back to a server,
 -- and validating that the server output matches up with another log.
-module Language.Haskell.LSP.Test.Replay
+module Language.LSP.Test.Replay
   ( -- replaySession
   )
 where
@@ -10,8 +10,8 @@ import           Control.Concurrent
 import           Control.Monad.IO.Class
 import qualified Data.ByteString.Lazy.Char8    as B
 import qualified Data.Text                     as T
-import           Language.Haskell.LSP.Types
-import           Language.Haskell.LSP.Types.Lens as LSP
+import           Language.LSP.Types
+import           Language.LSP.Types.Lens as LSP
 import           Data.Aeson
 import           Data.Default
 import           Data.List
@@ -20,12 +20,12 @@ import           Control.Lens hiding (List)
 import           Control.Monad
 import           System.FilePath
 import           System.IO
-import           Language.Haskell.LSP.Test
-import           Language.Haskell.LSP.Test.Compat
-import           Language.Haskell.LSP.Test.Files
-import           Language.Haskell.LSP.Test.Decoding
-import           Language.Haskell.LSP.Test.Server
-import           Language.Haskell.LSP.Test.Session
+import           Language.LSP.Test
+import           Language.LSP.Test.Compat
+import           Language.LSP.Test.Files
+import           Language.LSP.Test.Decoding
+import           Language.LSP.Test.Server
+import           Language.LSP.Test.Session
 
 {-
 -- | Replays a captured client output and
